requirements.txt
python-dotenv
requests
py-cord==2.0.0b1
